CS50 Problem set
Caesar
Add code for caesar, which uses a numeric key to shift letters in a
message key number of indexes to encryped the plaintext to get a
cyphired text.

Cash

Add code for the cash problem set, which calculates the number of coins
needed for a given amount of cents owed. The greedy algorithm is used,
mmeaning the larger coint (quarter) is first used, then the next in line
and so forth, until the total cents owed is accounted for.